refactor of ResetPassword to make it more easily testable
